Add Mandelbrot set creation/visualisation files
Add mandelbrot_engine which produces an image of a given region of the
Mandelbrot set to a certain number of iterations. Uses numba and numpy
for acceleration.

Add mandelbrot_static, saving a given image of a region of the
Mandelbrot set as output.png. Uses mandelbrot_engine

Add mandelbrot_dynamic, an interactable process for users to explore
the mandelbrot set however they like